# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A distributed task queue API for enqueuing and monitoring jobs.
  title: Task Queue API
  version: "1.0"
paths:
...
  /api/v1/circuit-breakers/reset/{type}:
    post:
      description: Closes the circuit breaker for a plugin type.
      parameters:
      - description: Plugin type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset a circuit breaker
      tags:
      - circuit-breaker
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticates with operator credentials and starts a server-side
        session.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          properties:
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login
      tags:
      - auth
...
swagger: "2.0"